Implemented routes for user register and login
