+++ backend/index.js
@@ -3,7 +3,7 @@ import mongoose from 'mongoose';
 import cors from 'cors';
 import dotenv from 'dotenv';
 
-import populateRoute from './routes/populate.js';
+import authRoute from './routes/auth.js';
 
 dotenv.config({ path: '../.env' });
 
@@ -14,7 +14,7 @@ app.use(cors({
   credentials: true,
 }));
 app.use(express.json());
-app.use('/populate', populateRoute);
+app.use('/auth', authRoute);
 
 
 app.get('/', (req, res) => {
@@ -28,12 +28,3 @@ app.listen(8080, () => {
 mongoose.connect(process.env.MONGODB_URI)
   .then(() => console.log('Connected to MongoDB'))
   .catch(err => console.error(err));
-
-// Placeholder for POST /populate route
-// app.post('/populate', async (req, res) => {
-//   try {
-//     const result = await ...
-//   } catch (err) {
-//     res.status(500).json({ error: err.message });
-//   }
-// });
